+++ src/app/(management-portal)/patient/analytics/[id]/page.module.scss
