+++ src/app/(management-portal)/patient/analytics/layout.module.scss
