+++ src/app/(management-portal)/patient/analytics/page.tsx
