--- src/app/(management-portal)/patient/dashboard/page.module.css
@@ -1,17 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: row;
-  color: #8f9bba;
-}
-
-.navigation-panel {
-  display: flex;
-  flex-direction: column;
-  color: white;
-}
-
-.rest-of-page {
-  display: flex;
-  flex-direction: column;
-  color: #8f9bba;
-}
